[JSON-OPTIMIZATION] Revamp conversion from Model to JS objects.
We were using `JSON.parse(JSON.stringify)` to convert Model to a JS
object. This breaks occasionally when Node does not have enough memory
to parse the large objects.

This fixes the issue by moving away from using `JSON.parse`, and
referring back to the schema to get the valid attributes, and then
mapping the values back to these attributes as results.